Added notification

+++ album/views.py
@@ -1,5 +1,5 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from .models import Calendar, Image, Comment, Friend, ImageComment
+from .models import Calendar, Image, Comment, Friend, ImageComment, Notification
 from accounts.models import User
 from django.utils import timezone
 from datetime import datetime
@@ -11,76 +11,54 @@ from .forms import (
     ImageForm,
     ImageCommentForm,
 )
-
-from django.db.models import Count
+from mysite.settings import TEMPLATES
 
 class CalendarView(TemplateView):
     template_name = 'album/home.html'
 
+    # TEMPLATES[0]['OPTIONS']['context_processors'].insert(0, "album.context_processors.notification")
     def get(self, request):
-        current_year = datetime.today().year
-        current_month = datetime.today().month
+
+        # notifications = Notification.objects.filter(receiver=request.user).order_by('read', '-date')
+        #
+        # page = request.GET.get('page', 1)
+        # paginator = Paginator(notifications, 5)
+        # try:
+        #     notifications = paginator.page(page)
+        # except PageNotAnInteger:
+        #     notifications = paginator.page(1)
+        # except EmptyPage:
+        #     notifications = paginator.page(paginator.num_pages)
+
+
+
         posts = Calendar.objects.all().order_by('-date')
         thumbnail = Image.objects.all()
         form = CalendarForm(request.POST)
-        # posts = Calendar.objects.filter(date__year=current_year, date__month=current_month).order_by('date')
 
-
+        # 검색 코드
         title = request.GET.get('title', '')
+        description = request.GET.get('description', '')
         year_month_str = request.GET.get('year_month', '')
         emoticon = request.GET.get('emoticon', '')
-
-        if title and year_month_str or year_month_str:
+        if year_month_str:  # 날짜 검색했을 경우 -> year_month = 검색한 날짜
             year_month = datetime.strptime(year_month_str, "%Y-%m").date()
-            posts = posts.filter(date__year=year_month.year, date__month=year_month.month)
+        else:               # 날짜 검색안했을 경우 -> year_month = 현재 날짜
+            year_month = datetime.strptime((str(datetime.today().year) +"-"+ str(datetime.today().month)), "%Y-%m").date()
+        posts = posts.filter(date__year=year_month.year, date__month=year_month.month, emoticon__contains=emoticon, title__contains=title, description__contains=description)
 
-        elif year_month_str and emoticon or emoticon:
-            year_month = datetime.strptime((str(current_year) +"-"+ str(current_month)), "%Y-%m").date()
-            posts = posts.filter(emoticon__contains=emoticon)
-
-        elif title:
-            year_month = datetime.strptime((str(current_year) +"-"+ str(current_month)), "%Y-%m").date()
-            posts = posts.filter(title__contains=title)
-        # elif year_month_str:
-        #     year_month = datetime.strptime(year_month_str, "%Y-%m").date()
-        #     posts = posts.filter(date__year=year_month.year, date__month=year_month.month)
-
-        else:
-            year_month = datetime.strptime((str(current_year) +"-"+ str(current_month)), "%Y-%m").date()
-            posts = posts.filter(date__year=year_month.year, date__month=year_month.month)
-            next_year_month = datetime.strptime((str(current_year) +"-"+ str(current_month)), "%Y-%m").date().month + 1
-            pre_year_month = datetime.strptime((str(current_year) +"-"+ str(current_month)), "%Y-%m").date().month - 1
-
-        pre_year = ''
-        next_year = ''
-        if year_month.month == 1:
-            pre_year = year_month.year - 1
-            pre_month = year_month.month + 11
-            next_year = year_month.year
-            next_month = year_month.month + 1
-        elif year_month.month == 12:
-            pre_year = year_month.year
-            pre_month = year_month.month - 1
-            next_year = year_month.year + 1
-            next_month = year_month.month - 11
-        else:
-            next_year = year_month.year
-            next_month = year_month.month + 1
-            pre_year = year_month.year
-            pre_month = year_month.month - 1
-
-        next_year_month = str(next_year) + "-" + str(next_month)
-        pre_year_month = str(pre_year) + "-" + str(pre_month)
-
-
-        # 현재 로그한 유저는 유저목록에서 제외시켜 출력하기 위해
-        users = User.objects.exclude(id=request.user.id)
-        # friend = Friend.objects.get(current_user=request.user) => need to create
-        friend, created = Friend.objects.get_or_create(current_user=request.user)
-        friends = friend.users.all()
-
+        # 이전달 / 다음달 구하기
+        pre_month = (year_month.month - 1) % 12 or 12
+        next_month = (year_month.month + 1) % 12 or 12
+        next_year_month = str(year_month.year) + "-" + str(next_month)
+        pre_year_month = str(year_month.year) + "-" + str(pre_month)
+        if pre_month is 12:
+            pre_year_month = str(year_month.year - 1) + "-" + str(pre_month)
+        elif next_month is 1:
+            next_year_month = str(year_month.year + 1) + "-" + str(next_month)
 
         args = {
+            # 'notifications': notifications,
             'form': form,
             'year_month': year_month,
             'next_year_month': next_year_month,
@@ -89,8 +67,6 @@ class CalendarView(TemplateView):
             'emoticon': emoticon,
             'posts': posts,
             'thumbnail': thumbnail,
-            'users': users,
-            'friends': friends,
         }
         return render(request, self.template_name, args)
 
@@ -125,34 +101,39 @@ def post_new(request):
 @login_required
 def post_edit(request, pk):
     post = get_object_or_404(Calendar, pk=pk)
-    images = Image.objects.filter(post=post)
+    if request.user.is_superuser or request.user == post.author:
+        images = Image.objects.filter(post=post)
 
-    if request.method == "POST":
-        form = CalendarForm(request.POST, instance=post)
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.author = request.user
-            post.save()
+        if request.method == "POST":
+            form = CalendarForm(request.POST, instance=post)
+            if form.is_valid():
+                post = form.save(commit=False)
+                post.author = request.user
+                post.save()
 
-            for img in request.FILES.getlist('image'):
-                Image.objects.create(
-                    post=post,
-                    image=img
-                ).save()
+                for img in request.FILES.getlist('image'):
+                    Image.objects.create(
+                        post=post,
+                        image=img
+                    ).save()
 
-            return redirect('album:post_detail', pk=post.pk)
+                return redirect('album:post_detail', pk=post.pk)
+        else:
+            form = CalendarForm(instance=post)
+        args = {
+            'form': form,
+            'images': images,
+        }
+        return render(request, 'album/post_edit.html', args)
     else:
-        form = CalendarForm(instance=post)
-    args = {
-        'form': form,
-        'images': images,
-    }
-    return render(request, 'album/post_edit.html', args)
+        return redirect('album:home')
+
 
 @login_required
 def post_remove(request, pk):
     post = get_object_or_404(Calendar, pk=pk)
-    post.delete()
+    if request.user.is_superuser or post.author == request.user:
+        post.delete()
     return redirect('album:home')
 
 class PostDetailView(TemplateView):
@@ -167,17 +148,23 @@ class PostDetailView(TemplateView):
         return render(request, self.template_name, args)
 
     def post(self, request, pk):
-        post_pk = get_object_or_404(Calendar, pk=pk)
+        post = get_object_or_404(Calendar, pk=pk)
         form = CommentForm(request.POST)
         if form.is_valid():
             comment = form.save(commit=False)
             comment.author = request.user
-            comment.post = post_pk
+            comment.post = post
             comment.save()
+            if post.author != request.user:
+                Notification.objects.create(
+                    receiver=post.author,
+                    sender=request.user,
+                    post_comment=comment,
+                ).save()
             form = CommentForm() # 입력후 다시 빈칸으로 만들기
-            return redirect('album:post_detail', pk=post_pk.pk)
+            return redirect('album:post_detail', pk=post.pk)
 
-        args =  {'form': form, 'post': post_pk}
+        args =  {'form': form, 'post': post}
         return render(request, self.template_name, args)
 
 class ImageDetailView(TemplateView):
@@ -215,6 +202,12 @@ class ImageDetailView(TemplateView):
             comment.image = image
             comment.author = request.user
             comment.save()
+            if image.post.author != request.user:
+                Notification.objects.create(
+                    receiver=image.post.author,
+                    sender=request.user,
+                    image_comment=comment,
+                ).save()
             form = ImageCommentForm() # 입력후 다시 빈칸으로 만들기
             return redirect('album:image_detail', post_pk=image.post.pk, pk=image.pk)
 
@@ -224,27 +217,30 @@ class ImageDetailView(TemplateView):
 @login_required
 def image_edit(request, pk):
     image = get_object_or_404(Image, pk=pk)
+    if request.user.is_superuser or request.user == image.post.author:
+        if request.method == "POST":
+            form = ImageForm(request.POST, instance=image)
+            if form.is_valid():
+                image = form.save(commit=False)
+                if request.FILES:
+                    image.image = request.FILES['image']
+                image.save()
 
-    if request.method == "POST":
-        form = ImageForm(request.POST, instance=image)
-        if form.is_valid():
-            image = form.save(commit=False)
-            if request.FILES:
-                image.image = request.FILES['image']
-            image.save()
-
-            return redirect('album:image_detail', post_pk=image.post.pk, pk=image.pk)
+                return redirect('album:image_detail', post_pk=image.post.pk, pk=image.pk)
+        else:
+            form = ImageForm(instance=image)
+        args = {
+            'form': form,
+        }
+        return render(request, 'album/image_edit.html', args)
     else:
-        form = ImageForm(instance=image)
-    args = {
-        'form': form,
-    }
-    return render(request, 'album/image_edit.html', args)
+        return redirect('album:home')
 
 @login_required
 def image_remove(request, operation, pk):
     image = get_object_or_404(Image, pk=pk)
-    image.delete()
+    if request.user.is_superuser or image.post.author == request.user:
+        image.delete()
     if operation == 'post_edit':
         return redirect('album:post_edit', pk=image.post.pk)
     elif operation == 'image_detail':
@@ -253,14 +249,27 @@ def image_remove(request, operation, pk):
 @login_required
 def comment_remove(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
-    comment.delete()
+    if request.user.is_superuser or comment.author == request.user:
+        comment.delete()
     return redirect('album:post_detail', pk=comment.post.pk)
 
 @login_required
 def image_comment_remove(request, pk):
     comment = get_object_or_404(ImageComment, pk=pk)
-    comment.delete()
+    if request.user.is_superuser or comment.author == request.user:
+        comment.delete()
     return redirect('album:image_detail', post_pk=comment.image.post.pk, pk=comment.image.pk)
+
+@login_required
+def notification(request, operation, pk):
+    if operation == 'post_comment':
+        comment = get_object_or_404(Comment, pk=pk)
+        notification = Notification.objects.filter(post_comment=comment).update(read=True)
+        return redirect('album:post_detail', pk=comment.post.pk)
+    elif operation == 'image_comment':
+        image_comment = get_object_or_404(ImageComment, pk=pk)
+        notification = Notification.objects.filter(image_comment=image_comment).update(read=True)
+        return redirect('album:image_detail', post_pk=image_comment.image.post.pk, pk=image_comment.image.pk)
 
 @login_required
 def change_friends(request, operation, pk):
